+++ packages/pyramid_lint/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Assist: `convert_to_for_in_iterable_indexed_loop`.
+
 ## [2.0.3] - 2024-03-15
 
 ### Changed
@@ -225,6 +231,7 @@
 [imsamgarg]: https://github.com/imsamgarg
 [jogboms]: https://github.com/jogboms
 
+[Unreleased]: https://github.com/charlescyt/pyramid_lint/compare/v2.0.3...HEAD
 [2.0.3]: https://github.com/charlescyt/pyramid_lint/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/charlescyt/pyramid_lint/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/charlescyt/pyramid_lint/compare/v2.0.0...v2.0.1